issue #6
This fixes the issue of using @Required on radio-button groups. The validator checks to see if the element is of type "radio", and if it has a "name" attribute. It then looks for all input elements that match have the same name and checks to see if any of those are checked. What this means is that in a radio-button group it is only necessary to attach @Required to one of the elements. Attaching @Required to more than one will return more than one constraint violation for the radio-button group since constraints are attached on a per-element basis.