issue #7
Part 1 of implementing Asynchronous Validation. There are a few TODO's
but I have started on restructuring the code to distinguish between
synchronous and asynchronous validation. Part of this refactor was to
ensure that current behavior remains unchanged. The next step will be to
implement actual asynchronous validation. There are numerous other
concerns, such as the impact on compound constraints and such. I have
annotated such concerns with TODO's.