issue #7
There was an issue to fix. Consider a compound constraint that contains
another compound constraint. Assume that we override the second
constraint such that it contains an asynchronous constraint. This means
that the parent constraint must also be marked async. Prior to this
commit, this was not the case. This commit should fix the issue. I did
this by adding a "parent" attribute to a CompositionGraph node and then
walking up the tree until the root node.