issue #7
Finished most of the async tests. I still have to write tests for
async-validation of form-specific constraints and radio groups. I
changed the behavior of regula#validate with respect to throwing errors.
In short, errors about constraints not being found will only be thrown
if elements are specified in conjunction with groups, constraints, or
groups and constraints. So calling regula.validate on just groups,
constraints, or groups and constraints which results in a situation
where there are no elements bound, will not result in an exception.