issue #7
Added tests for async testing of form-specific constraints and
radio-groups. This takes care of all testing necessary for asynchronous
constraints. At some point I should probably look at getting the
branch-coverage up. Luckily it is at 87%. Also updated the version of
JSCover.

This commit is also the last one for the asynchronous-validation
feature.