issue #36
Exposed validators through regula.Validator. For example, @Checked can
be accessed by regula.Validator.checked(element, params). The first
parameter will always be the element that you intend to validate. The
second parameter is an associative-array of parameters to the
constraint. The third parameter is a callback that must be provided if it
is an asynchronous constraint.

I may revisit this as I think more about it; this is only a first shot.